Move area-related actions to admin
